+++ lib/src/features/authentication/data/auth_actions_provider.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:todo_app/src/features/authentication/data/auth_repository_impl.dart';
+import 'package:todo_app/src/features/authentication/data/auth_repository.dart';
 import 'package:todo_app/src/features/authentication/data/auth_provider.dart';
 import 'package:todo_app/src/extensions/messages/auth_error_message.dart';
