+++ lib/src/features/authentication/data/auth_provider.dart
@@ -1,6 +1,6 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:todo_app/src/features/authentication/data/auth_repository_impl.dart';
+import 'package:todo_app/src/features/authentication/data/auth_repository.dart';
 
 // 現在の認証状態を確認するストリームプロバイダー
 final authStateProvider = StreamProvider<User?>((ref) {
